build(deps): update darling requirements from 0.21 to 0.23
Updates the requirements on [darling](https://github.com/TedDriggs/darling) to permit the latest
version.
- [Release notes](https://github.com/TedDriggs/darling/releases)
- [Changelog](https://github.com/TedDriggs/darling/blob/master/CHANGELOG.md)
- [Commits](https://github.com/TedDriggs/darling/compare/v0.21.0...v0.23.0)

---
updated-dependencies:
- dependency-name: darling
  dependency-version: 0.23.0
  dependency-type: direct:production
...

+++ qinterface/src/iface/global.rs
@@ -52,10 +52,10 @@ impl QuicInterfaces {
         factory: Arc<dyn ProductQuicIO>,
     ) -> BindInterface {
         let entry = self.interfaces.entry(bind_uri.clone());
-        if let Entry::Occupied(entry) = &entry {
-            if let Some(iface) = entry.get().iface().upgrade() {
-                return iface.binding();
-            }
+        if let Entry::Occupied(entry) = &entry
+            && let Some(iface) = entry.get().iface().upgrade()
+        {
+            return iface.binding();
         }
 
         let iface = Arc::new(RwInterface::new(bind_uri, factory, self.clone()));
@@ -101,12 +101,12 @@ impl QuicInterfaces {
 impl Interface {
     pub(super) fn close(&mut self) {
         self.io = Err(io::ErrorKind::NotConnected.into());
-        if let Entry::Occupied(entry) = self.ifaces.interfaces.entry(self.bind_uri.clone()) {
-            if entry.get().iface().upgrade().is_none() {
-                // NOTE: QuicInterfaces and Locations must be kept in sync.
-                Locations::global().remove_all(&self.bind_uri);
-                entry.remove();
-            }
+        if let Entry::Occupied(entry) = self.ifaces.interfaces.entry(self.bind_uri.clone())
+            && entry.get().iface().upgrade().is_none()
+        {
+            // NOTE: QuicInterfaces and Locations must be kept in sync.
+            Locations::global().remove_all(&self.bind_uri);
+            entry.remove();
         }
     }
 }
